requirements.txt
pandas
termcolor
iPython
cassandra-driver
